Fix sending the same change twice
When a line toggles faster than the usb message can be requested, only
the first change is storred. All subsequent changes op to when the usb
message is sent will be ignored.
This fixes the case where we read a change, ignore the next change, and
then recorde the change after that. This case would cause two usb
messages showing the same change occured twice.

Signed-off-by: Joseph A. Lutz <joseph.lutz@novatechweb.com>