Change the USB message API
Added to the functionality of the outputs. This requires changing how
the controll messages work. You can no longer change all the outputs at
once. They must be changed one at a time. The relay enable/disable has
it's own controll message. The wIndex is used to indicate which outputs
is being changed. The wValue is no longer a bitmast of the outputs state
but indicates weather to change the output on(!=0) or off(==0). The
output change control message must also contain 4 bytes of data. The
data is a delay time inmilliseconds. After the output state is changed a
timer is set to delay. After the delay expires the output will change to
to the opposite state of what was sent in the message.

Signed-off-by: Joseph A. Lutz <joseph.lutz@novatechweb.com>